# ism_swagger_RecyclEasy_218598.yaml
swagger: "2.0"
info:
  description: "The 'RecyclEasy' is a web application, that helps people find places near them, where they can dispose recyclabe garbage. As well as, publish advertisements related to recyclable garbage managment and disposal."
  version: "1.0.0"
  title: "RecycleEasy"
schemes:
  - "https"
  - "http"
paths:
...
  /changePassword:
    put:
      tags:
        - "auth"
      summary: "Change user password"
      description: ""
      operationId: "changePassword"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "header"
          type: string
          name: "Authorization"
          description: "Authorization header"
          required: true
        - in: "body"
          name: "body"
          description: "New password in duplicate"
          required: true
          schema:
            $ref: "#/definitions/ChangePasswordData"
      responses:
        200:
          description: "Password has been hanged"
        400:
          description: "Password is not strong. Passwords are not the same"
        403:
          description: "Only admin has access to this operation."
      security:
        - recycleasy_auth:
            - "write"
            - "read"
...
  /ad/create:
    post:
      tags:
        - "advertisement"
      summary: "Create advertisement"
      description: ""
      operationId: "advertisementCreate"
      consumes:
        - "application/json"
        - "application/x-www-form-urlencoded"
      produces:
        - "application/json"
      parameters:
        - in: "header"
          type: string
          name: "Authorization"
          description: "Authorization header"
          required: true
        - in: "body"
          name: "body"
          description: "Advertisement creation body"
          required: true
          schema:
            $ref: "#/definitions/AdvertisementCreationData"
      responses:
        200:
          description: "Advertisement create successfully."
        403:
          description: "Only admin has access to this operation."
      security:
        - recycleasy_auth:
            - "write"
            - "read"
  /ad/update/{advertisementId}:
    put:
      tags:
        - "advertisement"
      summary: "Edit advertisement with given id"
      description: ""
      operationId: "updateAdvertisement"
      consumes:
        - "application/json"
        - "application/x-www-form-urlencoded"
      produces:
        - "application/json"
      parameters:
        - in: "header"
          type: string
          name: "Authorization"
          description: "Authorization header"
          required: true
        - name: "advertisementId"
          in: "path"
          description: "Advertisement to modify."
          required: true
          type: number
      responses:
        200:
          description: "Advertisement updated successfully"
        403:
          description: "User not authorized."
        404:
          description: "Advertisement doesnt exist."
      security:
        - recycleasy_auth:
            - "write"
            - "read"
  /ad/delete/{advertsiementId}:
    delete:
      tags:
        - "advertisement"
      summary: "Delete advertisement with given id"
      description: ""
      operationId: "deleteAdvertisement"
      consumes:
        - "application/json"
        - "application/x-www-form-urlencoded"
      produces:
        - "application/json"
      parameters:
        - in: "header"
          type: string
          name: "Authorization"
          description: "Authorization header"
          required: true
        - name: "advertsiementId"
          in: "path"
          description: "Advertisement to delete."
          required: true
          type: number
      responses:
        200:
          description: "Advertisement deleted successfully"
        403:
          description: "User not authorized."
        404:
          description: "Advertisment doesnt exist."
      security:
        - recycleasy_auth:
            - "write"
            - "read"
...
  /ad/approve/{advertisementId}:
    post:
      tags:
        - "admin"
      summary: "Approve advertisement with given ID."
      description: ""
      operationId: "approveAd"
      consumes:
        - "application/json"
        - "application/x-www-form-urlencoded"
      produces:
        - "application/json"
      parameters:
        - in: "header"
          type: string
          name: "Authorization"
          description: "Authorization header"
          required: true
        - name: "advertisementId"
          in: "path"
          description: "ID of advertisement to approve"
          required: true
          type: number
      responses:
        200:
          description: "Advertisement approved successfully."
        403:
          description: "Only admin has access to this operation."
        404:
          description: "Advertisement doesnt exist."
      security:
        - recycleasy_auth:
            - "write"
            - "read"
  /ad/reject/{advertisementId}:
    post:
      tags:
        - "admin"
      summary: "Reject advertisement with given ID."
      description: ""
      operationId: "rejectAd"
      consumes:
        - "application/json"
        - "application/x-www-form-urlencoded"
      produces:
        - "application/json"
      parameters:
        - in: "header"
          type: string
          name: "Authorization"
          description: "Authorization header"
          required: true
        - name: "advertisementId"
          in: "path"
          description: "ID of advertisement to reject"
          required: true
          type: number
      responses:
        200:
          description: "Advertisement rejected successfully."
        403:
          description: "Only admin has access to this operation."
        404:
          description: "Advertisement doesnt exist."
      security:
        - recycleasy_auth:
            - "write"
            - "read"
securityDefinitions:
  recycleasy_auth:
    type: "oauth2"
    authorizationUrl: "http://find.your.team.com/login"
    flow: "implicit"
    scopes:
      write: "modify data"
      read: "read data"
  api_key:
    type: "apiKey"
    name: "api_key"
    in: "header"
definitions:
...
  ChangePasswordData:
    type: object
    required:
      - newPassword
      - password
    properties:
      password:
        type: string
      newPassword:
        type: string
...
  AdvertisementCreationData:
    type: object
    required:
      - title
      - description
      - offerType
      - phone
      - email
    properties:
      title:
        type: string
      description:
        type: string
      offerType:
        type: string
      email:
        type: string
      phone:
        type: string
...